V1.3

+++ lib/core/utils/validator.dart
@@ -4,7 +4,7 @@ class Validator {
     if (value == null || value.isEmpty) {
       return 'Please enter some text';
     }
-    if (double.tryParse(value) == null) {
+    if (int.tryParse(value) == null) {
       return 'Please enter a valid number';
     }
     return null;
